fix: address all Codacy static analysis issues
- Add void operator to floating promises in useConversationRunner tests
- Replace bracket access with slice() to fix object injection warning
- Remove unnecessary optional chain on model_id in RequestLogDetail
- Refactor StatusBadge with data-driven approach (CC 11 -> ~7)
- Remove truncate prop from DetailItem (9 -> 7 params)
- Extract ChatMessageList shared test helpers to reduce LOC
- Split useArenaState.state.test.ts into 3 focused files (578 NCL -> ~290 each)
- Extract AppLogDetail component from LogDetailModal (65 -> ~8 LOC)
