+++ web/src/pages/Chat/__tests__/ChatMessageList.interactions.test.tsx
@@ -1,80 +1,12 @@
 import { screen } from "@testing-library/react";
 import { renderWithProviders } from "../../../test/utils";
-import type { ChatMessage, Model } from "../../api/types";
 import { ChatMessageList } from "../ChatMessageList";
+import { createDefaultProps } from "./ChatMessageList.helpers";
 
 describe("ChatMessageList", () => {
-	const mockMessages: ChatMessage[] = [
-		{
-			role: "user",
-			content: "Hello, how are you?",
-			timestamp: Date.now(),
-		},
-		{
-			role: "assistant",
-			content: "I'm doing well, thank you!",
-			model: "Ollama Cloud/gemma3:4b",
-			timestamp: Date.now() + 1000,
-			metrics: {
-				durationMs: 500,
-				promptTokens: 10,
-				completionTokens: 20,
-				tokensPerSecond: 40,
-			},
-		},
-		{
-			role: "assistant",
-			content: "This is model B's response.",
-			model: "Ollama Cloud/glm-5",
-			timestamp: Date.now() + 2000,
-			metrics: {
-				durationMs: 600,
-				promptTokens: 12,
-				completionTokens: 25,
-				tokensPerSecond: 41.67,
-			},
-		},
-	];
-
-	const defaultProps = {
-		messages: mockMessages,
-		chatSubMode: "chat" as const,
-		isStreaming: false,
-		selectedModelB: "Ollama Cloud/glm-5",
-		enabledModels: [
-			{
-				id: "model-1",
-				model_id: "gemma3:4b",
-				name: "gemma3:4b",
-				description: "",
-				display_name: "gemma3:4b",
-				provider_id: "provider-1",
-				provider_name: "Ollama Cloud",
-				capabilities: '{"vision":false,"audio_input":false,"reasoning":false}',
-				params: "{}",
-				modality: "text",
-				input_modalities: "text",
-				output_modalities: "text",
-				context_length: 8192,
-				max_output_tokens: 4096,
-				input_price_per_million: null,
-				input_price_per_million_cache_hit: null,
-				output_price_per_million: null,
-				owned_by: "ollama",
-				enabled: true,
-				disabled_manually: false,
-				created_at: "2024-01-01T00:00:00Z",
-				last_seen_at: "2024-01-01T00:00:00Z",
-			},
-		] as Model[],
-		onStopConversation: vi.fn(),
-		onStop: vi.fn(),
-		onRegenerate: vi.fn(),
-		onDeleteMessage: vi.fn(),
-		activePersonaIdB: null,
-		conversationActivePersonaIdA: null,
-		chatActivePersonaId: null,
-	};
+	const defaultProps = createDefaultProps({
+		enabledModels: [createDefaultProps().enabledModels[0]],
+	});
 
 	beforeEach(() => {
 		vi.clearAllMocks();
