+++ web/src/pages/Chat/__tests__/useConversationRunner.edge-cases.test.ts
@@ -150,7 +150,7 @@ describe("useConversationRunner", () => {
 		});
 
 		act(() => {
-			result.current.runConversation();
+			void result.current.runConversation();
 		});
 
 		expect(params.setConversationState).not.toHaveBeenCalled();
