+++ web/src/pages/Chat/__tests__/useConversationRunner.messages.test.ts
@@ -71,7 +71,7 @@ describe("useConversationRunner", () => {
 		expect(promptBCount).toBeGreaterThanOrEqual(2);
 		// Verify ABAB interleaving: consecutive pairs should contain one A and one B
 		for (let i = 0; i < calledSystemPrompts.length - 1; i += 2) {
-			const pair = [calledSystemPrompts[i], calledSystemPrompts[i + 1]];
+			const pair = calledSystemPrompts.slice(i, i + 2);
 			const hasA = pair.some((p) => p === "System prompt A");
 			const hasB = pair.some((p) => p === "System prompt B");
 			expect(hasA && hasB).toBe(true);
